refactor: split settings_screen into display and emulators sections, fix analyze warnings
