+++ lib/ui/screens/settings_emulators_section.dart
@@ -1,0 +1,158 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:file_picker/file_picker.dart';
+import '../../core/storage/directory_service.dart';
+import '../../core/emulator/emulator_registry_data.dart';
+import '../../core/emulator/strategy_registry.dart';
+import '../../providers/download_provider.dart';
+
+// Function to build the Emulators section
+Widget buildEmulatorsSection(
+  BuildContext context,
+  DirectoryService directoryService,
+  bool emulatorsLoaded,
+  Map<String, bool> emulatorInstallStates,
+  Function(void Function()) setState,
+  WidgetRef ref,
+) {
+  return Column(
+    crossAxisAlignment: CrossAxisAlignment.start,
+    children: [
+      const Text('Emulators',
+          style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold)),
+      const SizedBox(height: 16),
+      if (!emulatorsLoaded)
+        const Center(child: CircularProgressIndicator())
+      else
+        ...kEmulatorDefinitions.map<Widget>((def) {
+          final emulatorId = def['id'] as String;
+          final emulatorName = def['name'] as String;
+          final isInstalled = emulatorInstallStates[emulatorId] ?? false;
+          final overridePath = directoryService.getEmulatorPathOverride(emulatorId);
+
+          return Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8.0),
+            child: Row(
+              children: [
+                Icon(
+                  isInstalled ? Icons.check_circle : Icons.cancel,
+                  color: isInstalled ? Colors.green : Colors.red,
+                ),
+                const SizedBox(width: 8),
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(emulatorName),
+                      if (overridePath != null)
+                        Text(
+                          overridePath,
+                          style: const TextStyle(fontSize: 11, color: Colors.grey),
+                          overflow: TextOverflow.ellipsis,
+                        ),
+                    ],
+                  ),
+                ),
+                IconButton(
+                  icon: const Icon(Icons.folder_open, size: 20),
+                  tooltip: 'Set custom directory',
+                  onPressed: () async {
+                    String? selectedDirectory = await FilePicker.platform.getDirectoryPath();
+                    if (selectedDirectory != null) {
+                      await directoryService.setEmulatorPathOverride(emulatorId, selectedDirectory);
+                      setState(() {}); // Trigger parent state update
+                    }
+                  },
+                ),
+                ElevatedButton(
+                  onPressed: isInstalled
+                      ? null
+                      : () async {
+                          // Capture context safely before any async operations.
+                          final messenger = ScaffoldMessenger.of(context);
+
+                          messenger.showSnackBar(SnackBar(
+                            content: Text('Starting download for $emulatorName...'),
+                          ));
+
+                          ref.read(downloadProvider.notifier).startEmulatorDownload(emulatorId, emulatorName);
+
+                          // Listen for download completion and update state
+                          ref.read(downloadProvider.notifier).stream.listen((downloads) {
+                            final progress = downloads[emulatorId];
+                            if (progress != null && progress.isComplete) { // Use safeContext here
+                              // Update the map directly. setState will re-render using the updated map.
+                              emulatorInstallStates[emulatorId] = true;
+                              setState(() {}); // Trigger parent state update
+                              messenger.showSnackBar(SnackBar(
+                                content: Text('$emulatorName downloaded.'),
+                              ));
+                            }
+                          });
+                        },
+                  child: Text(isInstalled ? 'Installed' : 'Download'),
+                ),
+              ],
+            ),
+          );
+        }),
+    ],
+  );
+}
+
+// Function to build the Emulator Conflicts section
+Widget buildConflictsSection(
+  StrategyRegistry registry,
+  Function(void Function()) setState,
+) {
+  final conflicts = registry.detectConflicts();
+  return Column(
+    crossAxisAlignment: CrossAxisAlignment.start,
+    children: [
+      const Text('Emulator Conflicts',
+          style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold)),
+      const SizedBox(height: 16),
+      if (conflicts.isEmpty)
+        const Text('No conflicts detected')
+      else
+        ...conflicts.entries.map((entry) {
+          final slug = entry.key;
+          final strategies = entry.value;
+          final currentStrategy = registry.getStrategyForSlug(slug);
+
+          return Padding(
+            padding: const EdgeInsets.symmetric(vertical: 8.0),
+            child: Row(
+              children: [
+                Expanded(
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      Text(slug, style: const TextStyle(fontWeight: FontWeight.w500)),
+                      Text(strategies.map((s) => s.name).join(' vs '),
+                          style: const TextStyle(fontSize: 12, color: Colors.grey)),
+                    ],
+                  ),
+                ),
+                DropdownButton<String>(
+                  value: currentStrategy?.emulatorId,
+                  items: strategies.map((s) {
+                    return DropdownMenuItem(
+                      value: s.emulatorId,
+                      child: Text(s.name),
+                    );
+                  }).toList(),
+                  onChanged: (value) async {
+                    if (value != null) {
+                      await registry.setPreference(slug, value);
+                      setState(() {}); // Trigger parent state update
+                    }
+                  },
+                ),
+              ],
+            ),
+          );
+        }),
+    ],
+  );
+}
